requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
playwright==1.52.0
pyee==13.0.0
typing_extensions==4.13.2
Werkzeug==3.1.3
psycopg2-binary
eng-to-ipa
